requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
django-filter
drf-yasg
celery[redis]
dj-config-url